requirements.txt
batchgenerators==0.25
einops==0.7.0
matplotlib==3.4.3
ml_collections==0.1.1
monai==1.3.0
nibabel==5.2.0
nni==2.9
numpy==1.24.1
opencv_python==4.8.1.78
Pillow==10.2.0
scipy==1.10.1
SimpleITK==2.3.1
Surface_Distance_Based_Measures==0.1
tifffile==2023.7.10
timm==0.9.12

torchtoolbox==0.1.8.2
torchvision
tqdm
